requirements.txt
botcity-framework-core>=1.0.0,<2.0
pandas

